app/src/main/java/com/auttmme/moviecatalogue/ui/movies/MovieViewModel.kt
package com.auttmme.moviecatalogue.ui.movies

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.auttmme.moviecatalogue.data.source.local.entity.MovieEntity
import com.auttmme.moviecatalogue.data.source.MovieCatalogueRepository

class MovieViewModel(private val movieCatalogueRepository: MovieCatalogueRepository) : ViewModel() {

    private var movieId: Int = 0

    fun setSelectedMovie(movieId: Int) {
        this.movieId = movieId
    }

    fun getMovie(): LiveData<MovieEntity> = movieCatalogueRepository.getMovieById(movieId)

    fun getAllMovies(): LiveData<List<MovieEntity>> = movieCatalogueRepository.getAllMovies()
}